Added Terminal
!!Please add the tcdlib.jar file to your eclips project to make this
work. I couldn't find a way to commit the added jar file.
Router and Users have terminals (with their names)
To send a message from a User type "send" into a user terminal it well
then ask for the message and for the name of the receipient.
!IMPORTANT!! when typing the name use caps lock for capital letters
using shift will add a question mark infront: ?Bob and the it won't
match the actual name